Add common ajax code
common/ajax.js contains code that both the node and browser version need.

To allow the common code to be flexible, processXmlResponse now expects an object with a single root object, rather than an object with type and data properties. Thus, xml2object now returns an object in the aforementioned format.

xhrRequestXml now returns an object with the processed xml as json and the xhr object.